实现了回复列表功能 当我们回复评论的时候可以在列表中快速显示，完善了问题的页面，通过问题页面简单的优化了下，同时在提交问题的时候 增加了校验（不能为空）。校验完成之后倒序排列问题的回复时间

+++ src/main/java/com/ccut/bbs/dto/QuestionDTO.java
@@ -12,8 +12,8 @@ public class QuestionDTO {
     private Long gmtCreate;
     private Long gmtModified;
     private Long creator;
-    private Long viewCount;
-    private Long commentCount;
-    private Long likeCount;
+    private Integer viewCount;
+    private Integer commentCount;
+    private Integer likeCount;
     private User user;
 }
